Refactor and improve code quality across multiple modules
- Simplified return statement in VM::run method.
- Added Default trait implementation for PluginRegistry.
- Changed function signature of values_to_json_array to accept a slice instead of a Vec.
- Updated test assertions for clarity and consistency.
- Removed unused test file import_as_star.rs.
- Reorganized imports in lowering_calls.rs and lowering_loop.rs for better readability.
- Enhanced readability of match statements in lowering tests.
- Improved error handling and assertions in plugin runtime integration tests.
- Fixed floating-point comparison in value roundtrip tests.
- Streamlined glob and exec functions in stdlib_plugin for better readability.
- Added safety documentation for FFI functions in stdlib_plugin.
- Refactored array handling in typed.rs for clarity and efficiency.
- Updated argument parsing in util.rs to use first() for consistency.

+++ cli/src/disassembler.rs
@@ -129,7 +129,7 @@ pub fn disassemble(bytes: &[u8]) -> Result<String, String> {
                 }
             }
             IROp::Jump { target } => {
-                let t = *target as usize;
+                let t = *target;
                 if let Some(name) = label_map.get(&t) {
                     out.push_str(&format!("{:04}  Jump {} ({})\n", i, name, t));
                 } else if let Some(name) = label_ordinals.get(t) {
@@ -139,7 +139,7 @@ pub fn disassemble(bytes: &[u8]) -> Result<String, String> {
                 }
             }
             IROp::BrTrue { cond, target } => {
-                let t = *target as usize;
+                let t = *target;
                 if let Some(name) = label_map.get(&t) {
                     out.push_str(&format!("{:04}  BrTrue r{} -> {} ({})\n", i, cond, name, t));
                 } else if let Some(name) = label_ordinals.get(t) {
@@ -152,7 +152,7 @@ pub fn disassemble(bytes: &[u8]) -> Result<String, String> {
                 }
             }
             IROp::BrFalse { cond, target } => {
-                let t = *target as usize;
+                let t = *target;
                 if let Some(name) = label_map.get(&t) {
                     out.push_str(&format!(
                         "{:04}  BrFalse r{} -> {} ({})\n",
@@ -221,7 +221,7 @@ pub fn disassemble(bytes: &[u8]) -> Result<String, String> {
                 label_index,
                 args,
             } => {
-                let t = *label_index as usize;
+                let t = *label_index;
                 if let Some(name) = label_map.get(&t) {
                     out.push_str(&format!("{:04}  CallLabel r{} <- {}(", i, dest, name));
                 } else {
